everything put on slam

+++ src/perception_sensor_lib/include/perception_sensor_lib/loop_closure/lap_counter.hpp
@@ -28,7 +28,7 @@ public:
     const Eigen::Vector3d& current_pose,
     const std::vector<Eigen::Vector2d>& map_cones,
     const Eigen::VectorXi& associations,
-    const Eigen::VectorXi& observations) const override;
+    const Eigen::VectorXd& observations) const override;
 
 private:
   double threshold_dist_;
